.claude/make-it/scaffolds/fastapi-nextjs/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
PyJWT
httpx
python-multipart
